% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprs.test}
\alias{exprs.test}
\title{A dataset for testing the function}
\format{
\subsection{\code{exprs.test}}{

A data frame with 37403 rows (genes) and 186 columns (samples):
\describe{
\item{country}{Netherlands}
\item{year}{2022}
...
}
}
}
\source{
AMC
}
\usage{
exprs.test
}
\description{
A dataset for testing the function
}
\keyword{datasets}
