% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseCMS-package.R
\docType{package}
\name{mouseCMS-package}
\alias{mouseCMS}
\alias{mouseCMS-package}
\title{mouseCMS: CMS classifier for mouse models}
\description{
A package to classify mouse models into CMS subtypes.This package contains a function for stratifying mouse models.
}
\author{
\strong{Maintainer}: Arezo Torang \email{torang.arezo@gmail.com} (\href{https://orcid.org/0000-0002-7912-3964}{ORCID})

}
\keyword{internal}
