% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseCMSclassifier.R
\name{mouseCMSclassifier}
\alias{mouseCMSclassifier}
\title{mouseCMSclassifier}
\usage{
mouseCMSclassifier(data, perform.log2 = FALSE)
}
\arguments{
\item{data}{A numeric matrix or data.frame with expression levels. Rows are genes and columns are samples.
Ensembl IDs are used.}

\item{perform.log2}{If \code{perform.log2 = TRUE}, the data will be log2-transformed.
It is a logical determining if data needs log2-transformation or if the data is already log2-transformed.
Default is \code{perform.log2 = FALSE}.}
}
\value{
An object of class "\code{data.frame}" containing the classification results.
Columns CMS1-4 show the chance of being each subtype, based on an elastic-net model.
nearestCMS indicates subtype with highest chance, regardless of the cutoff.
predictedCMS, shows subtype of samples that confidently predicted (cuttoff = 0.5).
}
\description{
\code{mouseCMSclassifier} is a CMS (Consensus Modelcular Subtypes in colorectal cancer)
classifier for mouse organoids.
}
\examples{
#classify a predefined set
data("exprs.test")
mouseCMSclassifier(data=exprs.test, perform.log2=FALSE)


}
